ref: separate EntityCustomPlayerMovement into existing classes
