+++ src/main/java/cam72cam/immersiverailroading/entity/EntityMoveableRollingStock.java
@@ -27,7 +27,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public abstract class EntityMoveableRollingStock extends EntityCustomPlayerMovement implements ICollision {
+public abstract class EntityMoveableRollingStock extends EntityRidableRollingStock implements ICollision {
 
     public static final String DAMAGE_SOURCE_HIT = "immersiverailroading:hitByTrain";
     public static final String DAMAGE_SOURCE_HIT_IN_DARKNESS = "immersiverailroading:hitByTrainInDarkness";
