+++ src/main/java/cam72cam/immersiverailroading/util/VecUtil.java
@@ -3,25 +3,22 @@ package cam72cam.immersiverailroading.util;
 import cam72cam.mod.math.Vec3d;
 import cam72cam.mod.util.Axis;
 import cam72cam.mod.util.FastMath;
-import util.Matrix4;
 
 public class VecUtil {
-	private VecUtil() {
-		// Disable construction since java does not have static classes
-	}
-
 	public static Vec3d fromYaw(double distance, float yaw) {
 		return new Vec3d(Math.sin(Math.toRadians(yaw)) * distance, 0, Math.cos(Math.toRadians(yaw)) * distance);
 	}
+
 	public static float toYaw(Vec3d delta) {
 		float yaw = (float) Math.toDegrees(FastMath.atan2(delta.x, delta.z));
 		return (yaw + 360f) % 360f;
 	}
+
 	public static Vec3d rotateYaw(Vec3d pos, float rotationYaw) {
 		if (rotationYaw - 90 == 0) {
 			return pos;
 		}
-		//return new Matrix4().rotate(Math.toRadians(rotationYaw-90), 0, 1, 0).apply(pos);
+
 		double cos = Math.cos(Math.toRadians(rotationYaw - 90));
 		double sin = Math.sin(Math.toRadians(rotationYaw - 90));
 		return new Vec3d(
@@ -30,12 +27,17 @@ public class VecUtil {
 				-sin * pos.x + cos * pos.z
 		);
 	}
+
 	public static Vec3d rotatePitch(Vec3d pos, float rotationPitch) {
-		if (Math.abs(rotationPitch) < 0.01) {
+		if(Math.abs(rotationPitch) == 0) {
 			return pos;
 		}
-		// TODO optimize me!
-		return new Matrix4().rotate(Math.toRadians(rotationPitch), 0, 0, 1).apply(pos);
+
+		double cos = Math.cos(Math.toRadians(rotationPitch));
+		double sin = Math.sin(Math.toRadians(rotationPitch));
+		return new Vec3d(pos.x,
+						 pos.y * cos + pos.z * sin,
+						 pos.z * cos - pos.y * sin);
 	}
 
 	public static Vec3d fromWrongYaw(double distance, float yaw)  {
